+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.8] - 2024-01-25
 ### Added
 - Support for `login_hint` query parameter to `GET /login` endpoint.
 - Support for `identifier` query parameter to `GET /login/identifier` endpoint.
@@ -81,8 +83,9 @@ it will redirect to `/login/password` as before.
 
 - Initial release.
 
-[Unreleased]: https://github.com/authnomicon/login/compare/v0.0.7...HEAD
-[0.0.6]: https://github.com/authnomicon/login/compare/v0.0.6...v0.0.7
+[Unreleased]: https://github.com/authnomicon/login/compare/v0.0.8...HEAD
+[0.0.8]: https://github.com/authnomicon/login/compare/v0.0.7...v0.0.8
+[0.0.7]: https://github.com/authnomicon/login/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/authnomicon/login/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/authnomicon/login/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/authnomicon/login/compare/v0.0.3...v0.0.4